Update README.md
Semi-supervised-clustering is a collection of state-of-the-art (SOTA),novel Semi-supervised-clustering methods (papers, codes and datasets). Any other interesting papers and codes are welcome. Any problems, please contact 508358959@qq.com. If you find this repository useful to your research or work, it is really appreciated to star this repository. ✨ If you use our code or the processed datasets in this repository for your research, please cite 1-2 papers in the citation part here. ❤️